# Don't use AEQ in removeUnusedExpr