Before:
```
benchmarking normalization of tests/shouldwork/Basic/AES.hs
time                 373.0 ms   (NaN s .. 429.7 ms)
                     0.997 R²   (0.995 R² .. 1.000 R²)
mean                 354.8 ms   (349.6 ms .. 364.7 ms)
std dev              9.762 ms   (312.1 μs .. 11.60 ms)
variance introduced by outliers: 19% (moderately inflated)
```

after:
```
benchmarking normalization of tests/shouldwork/Basic/AES.hs
time                 323.7 ms   (306.5 ms .. 355.1 ms)
                     0.996 R²   (0.985 R² .. 1.000 R²)
mean                 319.7 ms   (315.7 ms .. 327.6 ms)
std dev              8.781 ms   (93.23 μs .. 10.84 ms)
variance introduced by outliers: 16% (moderately inflated)
```